Add a setting for connection reuse
Before, the polling loop called "connect" each time through, even if there
was still an open/existing TCP connection. This commit fixes that bug by
not reconnecting, and instead adding an option to force disconnect and
reconnect (for those times you _know_ it's going to fail/time out).

Because it's impossible to tell if the other side closed the TCP
connection without sending data, and the API of PySunspec does not include
a way to automatically retry a connection when it fails, this seems like
the most sensible workaround.

The only times a loop through the polling thread does not return any data
now is when the connection is set to reuse an existing connection and that
connection went way (timeout, reset by peer, etc.)+